Migrate flight retrieval to the web API
The info provided is more helpful than the mobile API. It tells you when
a flight has departed, checked in, etc. I'll explore these attributes
after the initial migration.

The new flight info doesn't have the airport names, just the codes. I
parsed the list from Southwest's website and included it as a JSON file
mapping each airport code to the airport name and timezone. This will be
better overall because it isn't from a third-party database, but
directly from Southwest.